ui: Follow Figma in adding app-bar bottom border
The default app-bar color is now extremely similar to the scaffold
background color -- Color(0xfff5f5f5) and Color(0xfff6f6f6),
respectively -- and both follow the Figma. Because of their
similarity, it seems helpful to delineate the app bar a bit better,
and we can do so by adding this bottom border, which the Figma has:
  https://www.figma.com/file/1JTNtYo9memgW7vV6d0ygq/Zulip-Mobile?type=design&node-id=143-10772&mode=design&t=Nvsl6UdRBQvH3zXV-0

In the message list, the Figma isn't consistent on whether to omit
the border when the app bar matches the color of all the recipient
headers, like in a stream narrow. I think it's better to omit it in
those cases, so I've done that here.