android [nfc]: Cut duplicate INTERNET permission from profile manifest
In the initial files that `flutter create` gave us from the template,
both the profile and debug manifests claimed this permission.  Early
in the prototype, in 27ea3c8da, we claimed it in the main manifest
too, because this is an app that naturally needs it; and I removed
the duplicate copy in the debug manifest but didn't notice the profile
manifest at the time.  Cut it from there too.