content [nfc]: Set error{,Code}Style font size to kBaseFontSize explicitly
kBaseFontSize is currently 14.

Before this, the text was sized by
Theme.of(context).textTheme.bodyMedium.fontSize, thanks to an
AnimatedDefaultTextStyle in the Material widget. That value is 14
when the app's localization state is set to render text as
"English-like" (see [Typography]), which at this point I think it
always is.

We want to increase the content font size soon, for #512, and for
that it'll be helpful if everything uses kBaseFontSize or some
multiple of it. From scrolling through a bunch of content with
kBaseFontSize increased dramatically, I think these were the only
pieces of content that weren't doing so.