deps android: Upgrade to Android Gradle Plugin 8.1.4
This is the version that the "AGP Upgrade Assistant"
in Android Studio recommends as the next upgrade.
Release notes:

  https://developer.android.com/build/releases/past-releases/agp-8-0-0-release-notes
  https://developer.android.com/build/releases/past-releases/agp-8-1-0-release-notes

The three flags added to gradle.properties were put there
by the AGP Upgrade Assistant, and they each have the effect
of preserving some behavior where AGP 8.0 changed the default.
I'll go on to remove them in the next couple of commits, but
each with their own commit message.

The one other change mentioned in the notes that looks like it could
possibly affect us is that R8 now runs in "full mode" by default.
This seems to come down to more aggressive tree-shaking of JVM code.
We write very little such code ourselves, so the most likely way that
could have an effect is if some dependency of ours is doing something
fancy that breaks the assumptions that relies on.

The one place I'd be most concerned about some dependency ceasing to
work because some of its code was tree-shaken away is in receiving
notifications.  So I just went and manually tested notifications, in
debug and release builds, when the app is in the foreground, the
background, or not running.  All still work.