android: Stop generating BuildConfig class
This flag android.defaults.buildfeatures.buildconfig is false
by default since AGP 8.0.  It was set to true by the upgrade
assistant; switch it to false now.

AGP is, as always, thinly documented.  The AGP 8.0 release notes:
  https://developer.android.com/build/releases/past-releases/agp-8-0-0-release-notes

said this controls whether to generate `BuildConfig`.  For that,
here's some indirect documentation in a list of "Gradle tips":
  https://developer.android.com/build/gradle-tips#share-custom-fields-and-resource-values-with-your-app-code

and here's someone's Medium post which is also informative (even
though only the start of it is visible):
  https://blog.dipien.com/stop-generating-the-buildconfig-on-your-android-modules-7d82dd7f20f1

In short, eliminating this is another build-speed improvement,
which again should only be able to cause problems by doing so
at compile time.  Sure, sounds good.