text: Set 0 letter spacing throughout zulipTypography
This zeroes out the letter spacing in message content (#545), and
also everywhere else that uses styles from zulipTypography without
overriding the letter spacing.

It's worth mentioning that letter spacing is one of the parameters
that are carefully chosen for the 15 TextThemes that make up the
default Typography in Material 3:

  {display,headline,title,label,body} × {small,medium,large}

So this might be considered a regression...if the Material spec were
our primary or only source of truth for text styles. But it isn't:
we have a designer (Vlad), and this change brings us closer to his
specified design. See #545 and #548.

I also checked to see if the Material 3 default varies the letter
spacing among the three [ScriptCategory]ies (englishLike, dense, and
tall), which might be a reason to ask Vlad if we want to do the
same. But the letter-spacing values are the same across the
[ScriptCategory]ies, so no need for that.

After this, there's still more to do for #548: letter spacing
of 1% (i.e., 0.01 times `fontSize`) in buttons.

Fixes: #545
Fixes-partly: #548