check: pin flutter SDK version
Use direnv and flutter to enforce and validate a pinned version of flutter.

This approach uses `direnv` and `fvm` to setup the environment with the pinned
version of the repo.  When setup correctly, `direnv` will turn on `fvm`'s
pinned flutter version whenever entering the zulip-flutter repository
(directory).
The `.envrc` does the following:
- Use fvm's flutter version if it's available.
- Check that whether the flutter binary version matches what was pinned for the
repo.
- Let users know if there's a mismatch of version so that the user can take
action.

Other notes:
- Power users can:
  - use their checked out flutter sdk (either their own custom path, configured
    in .envrc, or the system default)
  - disable the check if they know what they're doing and want to use a custom
    flutter installation