notif: Switch to our own Pigeon-based plugin for showing a notification
This delivers the first major tranche of #351, switching from
package:flutter_local_notifications to our own thin API binding
based on Pigeon.  With this commit, we've converted over one of
the handful of different methods we use from that package,
and it's the most complex of them by a substantial margin.

Fixes-partly: #351