deps: Add a shim package for a Flutter plugin of our own
For the moment the new plugin is an empty shell, doing nothing.
We'll add functionality to it shortly.

The reason we need this somewhat complicated arrangement is so that
the `flutter` tool sees this plugin and includes it in some code it
generates, in order to cause the Flutter engine to automatically
initialize the plugin on startup.  For details, see the comment
this commit includes in the new shim package's pubspec.