msglist: Reserve space for "mark as read" button so messages don't shift up/down
This commit addresses the issue of the message list shifting downwards
when the "mark as read" button is pressed. The previous implementation
would return an empty widget when the button was pressed and cause the
message list to shift down. Now it takes a widget which will display
the button based on the opacity level and will disable the region the
button holds once the button is pressed, thus won't cause the message
list to shift down.

Fixes: #562