compose test: Check ID still included if same-name users deactivated
While autocomplete does not include deactivated users, there still
is a possibility that same-name users with one or more deactivated
gets mentioned, which will make the mentioned user ambiguous without
the id, this is explained in #451.

This adds a test case that asserts same-name users are disambiguated
while composing the message regardless of them being active.

Fixes part of #451