deps: Upgrade packages to latest
Done because it's the first step in our plan for #612, the "Missing
API declaration" warning from Apple.

This is the result of `tools/upgrade pod-major`. No breaking changes
requiring synchronous code changes were found.

Changelogs:
  https://pub.dev/packages/device_info_plus/changelog
  https://pub.dev/packages/file_picker/changelog
  https://pub.dev/packages/flutter_local_notifications/changelog
  https://pub.dev/packages/package_info_plus/changelog
  https://pub.dev/packages/share_plus/changelog
  https://pub.dev/packages/pigeon/changelog

All of these changelogs except `pigeon` mention adding "privacy
info" or "privacy manifest" in the version we're taking here --
exactly the kind of thing we want to see, with #612 in mind. And
`pigeon` shouldn't need one, as it's a dev dependency only, not part
of the actual build.

Related: #612