android: Bump minSdkVersion to 28, aka Android 9 Pie
In 2023-09, just 2.2% of our zulip-mobile Android users were below
this version, according to stats Greg posted:
  https://chat.zulip.org/#narrow/stream/48-mobile/topic/platform.20versions/near/1648299

So I think by now we can be comfortable making this the new minumum.

I'm not sure if this is actually strictly required by some
dependency upgrades we'll take later in this series of commits. But
Android 9 is what the office Android device is on, and the core team
doesn't have a physical device running a lower version. So we'll
validate the upcoming dependency upgrades experimentally, by running
on the office Android device. And with this bump to the threshold,
we can leave it at that.

Discussion:
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/flutter.3A.20Java.20versions.3F/near/1775991