ios: Proactively bump min iOS version to 14.0, matching zulip-mobile
We could almost certainly bump it to 15, based on the last
platform-version statistics Greg posted, and that was in 2023-09:
  https://chat.zulip.org/#narrow/stream/48-mobile/topic/platform.20versions/near/1648327

But there isn't specific code we want to write for iOS 15 features
right now, so might as well leave that for later.

In fact, I don't think there's a particularly strong need to bump it
to 14, at least at the moment. But there's no reason it needs to lag
behind zulip-mobile. And in case the lower threshold would be an
obstacle later on, it's nice to have it cleared away proactively
ahead of time.