deps: Upgrade packages to latest, namely package_info_plus, share_plus
This is the result of `tools/upgrade pod-major`, and some small
adjustments for a breaking API change in share_plus.

Changelogs:
  https://pub.dev/packages/package_info_plus/changelog
  https://pub.dev/packages/share_plus/changelog

Tested on my iPhone and on the office Android device. In particular,
I checked that sharing message content via the message action sheet
still worked, and I checked that the "About Zulip" page still shows
the "App version" field correctly.