text: Add proportionalLetterSpacing
To confirm that [TextStyle.letterSpacing] is taken literally as
logical pixels, as its dartdoc suggests it is, I rendered the text
"||" with an extreme [TextStyle.letterSpacing] value of 40. Then it
was easy to see that the letter spacing -- the horizontal distance
between the two "|" characters -- wasn't changing as I adjusted the
text-size slider in my iPhone settings.

In the tests, the `textScaleFactors` list is copied from
emoji_reaction_test. Perhaps there's a convenient place to define it
centrally.