text: Use proportionalLetterSpacing in the few existing places that want it
The letter spacings in these pieces of text hasn't been responding
to system font-size adjustments. Now they do.

There might not be any other instances of this problem in the app,
because we widely apply 0 for [TextStyle.letterSpacing] (since
0065954d), and when we do that, it shouldn't matter whether zero is
considered as logical pixels or as a proportion the surrounding text
size.