ide: Include .idea/zulip.iml and .idea/modules.xml
In Greg's testing, the modules.xml file seemed needed in order for
Android Studio to know where to find the .iml file.

This has some useful `excludeFolder` items that should help Android
Studio see that certain files are not part of our project, and
should be excluded in searches that are supposed to be limited to
"project files".