model [nfc]: Remove StreamColorSwatch; use ColorSwatch<StreamColorVariant>
This indirection was mildly helpful; it let us consume these colors
with slightly more concise code, and without needing `!` to satisfy
the analyzer.

But we get the same correct functionality without the indirection,
and a bit more transparently. The more substantive benefit of using
ColorSwatch directly is that we now have a ready-made lerp function
-- `ColorSwatch.lerp` -- that can operate on our swatch instances.
That should be helpful later when we want stream colors to animate
smoothly when dark mode is toggled, along with various other style
attributes that will naturally animate too.