message: Store all messages centrally
In itself, this exacerbates #455 by making it trigger in more
circumstances.  That's OK for the moment, because this also gives us
a foundation on which to fix that issue soon.

A key part of testing this is the line added in message_list_test's
`checkInvariants`, which checks that all messages tracked by the
MessageListView are present in the central message store.

Co-authored-by: Chris Bobbe <cbobbe@zulip.com>