api [nfc]: Rationalize "orig"/"new" field names on UpdateMessageEvent
In the server API here, "stream_id" is the old value and contrasts
with "new_stream_id", but "subject" is the new value and contrasts
with "orig_subject".

This makes them confusing to read, and is the way it is only for
historical reasons: I believe the story is that before Zulip
Server 3.0 added support for moving topics between streams, three
of those properties were already present and then "new_stream_id"
was squeezed in at the best name still available.

Fixing that in the API will be a pain because it'll need a migration
path for compatibility.  But we can fix it here within our code.