message: Mutate messages for message edits here, instead of in msglists
Like a recent commit did for reactions, this fixes the part of #455
that's concerned with message edits.

With message edits, unlike reactions, the double-processing issue
when there are two or more message lists wasn't user-facing, since
the mutation done by message lists was (naturally) idempotent. So
there isn't really an easy or helpful regression test to write for
that. But we can and do write one against the dropped-update symptom
when there are zero message lists open. For an explanation of that
symptom, see the recent commit where we dealt with reactions.

Co-authored-by: Chris Bobbe <cbobbe@zulip.com>
Fixes-partly: #455