content: Style @-mention pills with 2023+ direct-mention style background
As the TODO comments say, there's more we have to do to bring this
up-to-date with the new, 2023+ design.

But now at least we've removed a style that doesn't appear at all in
that new design, and we're left with something we can make a
dark-theme variant for by straightforwardly checking what web does.