autocomplete: Remove the `refreshStaleUserResults` method
This method was used to search for the new user results in autocomplete
when the users in store were modified during an in-progress computation.
Along with this, `_startSearch` would do the same thing by retrying,
resulting in duplicate actions. By removing this, the unnecessary search
for user results is prevented.

Besides, this commit also removes some live-updating behavior of
user-mention autocomplete results in order to provide a better UX.
To read more about this decision, see:
  https://github.com/zulip/zulip-flutter/pull/693#pullrequestreview-2091696195