msglist: Follow web in recipient-header-date color
The old color, which we started using for stream headers in
e97e6fd26 and DM headers in 81f70386, doesn't match the Figma:
  https://www.figma.com/design/1JTNtYo9memgW7vV6d0ygq/Zulip-Mobile?node-id=538-20849&t=ubXJTWC5kuz5awbN-0
  #666666
or the web app:
  --color-date,
  which is hsl(0deg 0% 15% / 75%)

I wonder if some confusion about contrastingColor (see a recent
commit where we removed a variable with that name) might have led to
the old color being used for stream headers, and perhaps the
DM-header change was just done to align with stream headers.

Anyway, might as well use the web app's color, which we were doing
before e97e6fd26. We can get a dark-theme variant from the web app
too, which is nice. The Figma doesn't yet offer a dark-theme
variant.