msglist: Follow web and Figma in recipient-header text color
In web this is known as --color-text-message-header,
and here is a Figma node:
  https://www.figma.com/design/1JTNtYo9memgW7vV6d0ygq/Zulip-Mobile?node-id=538-20849&t=O4Yz8kKlRrXFY8Te-0

This is actually the same color that web uses -- and we use -- for
message content. For that usage, web calls it
--color-text-message-default.

In fact, thanks to the DefaultTextStyle that wraps the whole message
list (for the sake of coloring message content), DM recipient header
text already had this color.

So one option would have been to just remove the black color on
stream recipient header text, to allow it to be colored via the same
inheritance. But I didn't want to do that. Web has a reason for
using separate variables: in dark theme, recipient header text is
colored slightly different from message content. So, we'll keep
these color definitions separate too, to make it easy to follow web
in dark theme.