ui: Adjust app bar and scaffold-background to align with new Figma design
From discussion:
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/design.3A.20colors/near/1801938

and the Figma linked there:
  https://www.figma.com/design/1JTNtYo9memgW7vV6d0ygq/Zulip-Mobile?node-id=2074%3A26325&t=rjuqcQaHMiGBUReF-1

Now, the app-bar bottom border is colored by the new Figma's
`border-bar`, and the main background color is `bg-main`. The app
bar background is unchanged because it already matches `bg-top-bar`.
I've labeled these values with comments in the code.

And all these color variables have dark-theme variants! (See
discussion for what those are.) That'll help when we implement dark
theme, #95.