recent_senders: Introduce `MessageIdTracker` and `RecentSenders` data structures
MessageIdTracker data structure is used to keep track of message ids in
an ascending sorted list. It is used in RecentSenders data structure.
RecentSenders data structure is used to keep track of user messages in topics
and streams.

Much of this code is transcribed from Zulip web; in particular,
from:
  https://github.com/zulip/zulip/blob/bd04a30bbc6dc5bd7c20940a3d1d34cf8c8c6f28/web/src/recent_senders.ts