msglist: Set message-sender-name color to match web
I overlooked this in #686 because this text's color wasn't set at
this site explicitly, and my process for that PR was to search
case-insensitively for the term [color]. Instead of being set
locally, this text's color was being set by the DefaultTextStyle
that wraps the whole message list for the sake of coloring message
content.

That color:
  const HSLColor.fromAHSL(1, 0, 0, 0.15).toColor()
doesn't match web, it turns out. (Or the old Figma, which uses
true black.) Web colors this with var(--color-text-sender-name), but
that variable is undefined for light theme, so it falls back to
--color-text-default which is this color.

At this commit, if I change the message list's DefaultTextStyle to
set a conspicuous color, like orange, I only see a response from
message content, and not from sender names or anything else. That's
good, but it would be better to scope the DefaultTextStyle to just
the message content, to avoid errors like this in the future. I'll
do that later in this series of commits.