content [nfc]: Scope content text color to just the content
The purpose of this DefaultTextStyle is to color message content.
Moving it leafward, so it proximally wraps message content, is
helpful to avoid the error of applying the color to other
message-list elements that don't want it. A recent commit fixed an
instance of this error with sender names. I believe there weren't
any other instances, so I've marked this NFC.