content: Follow web in setting 1.4x line height for code blocks
(And apply the same change to our simplified math blocks, which use
this variable too.)

This was ~1.29x in the previous commit, and 1.43x before that. (The
1.43x was conveniently very close to 1.4x, but that was basically an
accident. It's so close that on my device I can't see a difference
in a screenful of code block content...perhaps there's a step that
rounds to a physical-pixel count, and the difference might be
observable on a device with a higher pixel density than mine.)