deps [nfc]: Cut misplaced comment about gen-l10n
This isn't a helpful place for this information, because there's
no reason one would be looking in the pubspec at a time when one
needs it.  We have the same information about `flutter run` and
`flutter gen-l10n` in the README already, which is a better home.

The part about lib/l10n/ is just inaccurate; there's no such
directory in our tree.  I think that may have been the home of
what's now assets/l10n/, in a previous draft of the PR that
introduced this.