msglist test [nfc]: Make `prepare` helper async
This is the first in a series of steps toward making
[PerAccountStore.handleEvent] an async method.  We'll need that
in turn in order to let it start saving changes to disk when
applicable, like for #458, and having both it and its caller,
[UpdateMachine.poll], wait for the result rather than barreling
ahead to make other changes concurrently.

This change comes first because when we make `handleEvent` async,
we'll want to make sure all its call sites properly wait for it
to finish; and when that requires changing the caller to an async
method, we'll want to do the same for *its* call sites, recursively.
To keep the changes manageable to read and review, we'll make
separate commits for each method, except local helpers with only a
handful of call sites.  The callers have to go before the callees
in order to keep the commits coherent, so we start with this
transitive caller.