profile [nfc]: Wrap Paragraph widget in DefaultTextStyle
This is NFC because baseContentTextStyle just sets the color
attribute, and this particular Paragraph only renders a link, which
will clobber the default color with the link color.

But we're about to move some other text style attributes (size and
line height) from Paragraph to baseContentTextStyle, and we don't
want the link to lose those style attributes when we make that
change.