content [nfc]: Make textStylePlainParagraph complete; cut off inheritance
As Greg points out:

https://github.com/zulip/zulip-flutter/pull/698#issuecomment-2123861381
> Avoiding `DefaultTextStyle.merge`, in favor of giving some
> already-computed style directly, will be good for explicitness as
> well as for potentially a small performance gain.

Before this, the inheritance via `DefaultTextStyle.merge` made it
kind of tiring to work out what plain-paragraph styling actually
was. It was looking up to the nearest DefaultTextStyle, which was
actually an AnimatedDefaultTextStyle in the Material widget, that
was providing Theme.of(context).textTheme.bodyMedium. That, in turn,
was built from various things:
- [_M3Typography.englishLike]
- [Typography.blackCupertino] or [Typography.blackMountainView]
  (depending on platform)
- `zulipTypography` in lib/widgets/text.dart

I've followed these sources in writing down this complete style
object, to make this a direct, NFC translation.