deps: Upgrade packages to latest, namely pigeon, lints, flutter_lints
This is the result of `tools/upgrade pod-major`.

All of these are dev-only dependencies, so what you see is what you
get.  The only effect of the `pigeon` upgrade is the changes below
(which are only in comments noting the version itself); and given that
the `lints` and `flutter_lints` upgrades don't cause the analyzer to
complain, there's nothing else they could do to break things.

Looking at changelogs, the linter upgrades add and remove some lints
to the list, and the Pigeon upgrade has a breaking change for Swift.

Changelogs:
  https://pub.dev/packages/pigeon/changelog
  https://pub.dev/packages/lints/changelog
  https://pub.dev/packages/flutter_lints/changelog