message: Mutate messages for flags events here, instead of in msglists
Like a recent commit did for message edits, and one before that did
for reactions, this fixes the part of #455 that's concerned with
update-message-flags events.

Like in the commit that dealt with message edits, this fixes the
symptom of a dropped update when there are zero message lists open.
In this case I'm *pretty* sure there wasn't a user-facing symptom
with the double-processing situation when the message is in two or
more open message lists. That's because I believe we don't give
different behavior for a flag that's present multiple times in the
`flags` list, vs. just once.