api test [nfc]: Unfold an obscure deepEquals
The idea of the `(it) => it` here was that it's a condition callback
that imposes no conditions -- a no-op, just a placeholder for there
being some value there at all.

Here we can express the same outcome a lot more simply, though,
by forgetting about the values and looking at the keys.  So do that.