db test: Fix implicit dynamic in a condition callback, and tighten check
Like in the previous commit, this `(it) => it` was a no-op
condition callback.  Here, there isn't a way to just rearrange
the surrounding check to cut this out while simplifying.
So make the type explicit.

While we're at it, add just a bit more of a check.  This
demonstrates how to write such a callback, for use in `deepEquals`,
that expects a specific type of value.