nav [nfc]: Bound nav types to prevent implicit dynamic, toward strict-raw-types
We refer to these types like WidgetRoute in a lot of places in tests,
for expressions like:

    it.isA<WidgetRoute>().page.isA<ChooseAccountPage>()

Because we don't specify a type argument, the type parameters are
all getting inferred as `dynamic`.

We never actually want `dynamic`, and we want to tell the analyzer
to start complaining when it appears implicitly like that (with
`strict-raw-types`).  In these examples, the type parameter doesn't
even matter at all, though, so we'd like to avoid cluttering up the
tests with irrelevant type arguments (even `void`).  We can have our
way on both counts by making the bound `Object?`, so that the type
parameters default to that instead of to `dynamic`.