test [nfc]: Fix implicit dynamic in condition-callback types
These `it` parameters were implicitly getting typed `Subject<dynamic>`.
We want `Subject<Object?>`, as that's what `deepEquals` expects.
(The `dynamic` version was working fine, but would be flagged by the
analyzer under the strict-raw-types option.)

We could explicitly write that in each case... or we could tighten
a bit by using the `Condition` alias and applying it to the whole
lists that these appear in.