notif test [nfc]: Fix implicit dynamic Route type
And fix the explicit dynamic Route type on the related local variable
pushedRoutes.  The value hypothetically returned from these routes is
something we promise not to care about at all (i.e., type `void`),
not something we intend to work with but in a statically-untyped way
(i.e., type `dynamic`).