content [nfc]: Put code-block span styles in ContentTheme
With the code-block span styles all bundled together in a new
CodeBlockTextStyles class, it'll be convenient to make separate
`light` and `dark` constructors toward #95 dark theme.

Next, we'll pass a BuildContext to the CodeBlockTextStyles
constructor, in order to use weightVariableTextStyle for the spans
that are supposed to be bold. (The code-block font, Source Code Pro,
is variable-weight.) This would have been possible before the
current refactor, of course, but this way the spans don't have to
repeat the weightVariableTextStyle computations on-demand. They can
instead just look up the result from this central source.

Related: #95