content: Use weightVariableTextStyle for bold spans in code blocks
Now, as the added test confirms, bold spans are getting a `wght`
value of 700, instead of 400 from the ambient style.

The FontWeight.bold seems to have already been causing the spans to
render bolder than surrounding text, but in a way that doesn't match
what we get for Source Code Pro with `wght` 700. I think this is
because of
  https://github.com/flutter/flutter/issues/136779
.