content [nfc]: Reassert normal font weight for plain code-block text
I *think* it's not possible for a code block to appear in a bold
context. In particular I haven't found a way to put one inside a
spoiler header.

Anyway, just in case it does end up being possible somehow, might as
well clobber the bold with normal weight, to make sure the code
block's bold spans remain distinct from non-bold spans. It's
convenient (and cheap) to do that here in the CodeBlockTextStyles
constructor.