i18n: Follow stream-channel rename in our one translation, too
Before launch, in order to manage contributions from many people
for translations into many languages, we'll wire things up to
an appropriate platform: #276.  For now, though, we have just this
one partial translation into Japanese, as an end-to-end demo of the
i18n setup within the app, and we maintain it directly in the repo.
Update that translation to reflect this rename.