profile: Use weightVariableTextStyle instead of TextStyle.fontWeight
Now, this text is styled with a `wght` value of 700, instead of 400
from the ambient style. The FontWeight.bold seems to have already
been causing the spans to render more boldly than surrounding text,
but in a way that doesn't match what we get for Source Code Pro with
`wght` 700. I think this is because of
  https://github.com/flutter/flutter/issues/136779
.

The slightly changed appearance is making some profile field labels
run onto two lines on my device on CZO, instead of just one line, as
Alya pointed out:
  https://github.com/zulip/zulip-flutter/pull/728#issuecomment-2155514758
So I'll follow up by giving those labels a bit more horizontal room.